Extract composer runtime styles
