+++ index_editor.html
@@ -18,7 +18,7 @@
       } catch (_) { /* ignore */ }
     })();
   </script>
-  <link rel="stylesheet" href="assets/themes/native/theme.css?v=press-system-v3.4.46" data-editor-native-theme>
+  <link rel="stylesheet" href="assets/themes/native/theme.css?v=press-system-v3.4.47" data-editor-native-theme>
   <style>
     *, *::before, *::after { box-sizing: border-box; }
     html { height: 100%; }
@@ -2605,9 +2605,9 @@
       updateOffset();
     })();
   </script>
-  <script type="module" src="assets/js/editor-boot.js?v=press-system-v3.4.46"></script>
-  <script type="module" src="assets/js/editor-main.js?v=press-system-v3.4.46"></script>
-  <script type="module" src="assets/js/composer.js?v=press-system-v3.4.46"></script>
+  <script type="module" src="assets/js/editor-boot.js?v=press-system-v3.4.47"></script>
+  <script type="module" src="assets/js/editor-main.js?v=press-system-v3.4.47"></script>
+  <script type="module" src="assets/js/composer.js?v=press-system-v3.4.47"></script>
   <!-- Back to top button -->
   <button id="backToTop" class="back-to-top" aria-label="返回顶部" title="返回顶部" type="button" hidden>
     <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round" aria-hidden="true">
